requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
social-auth-app-django==5.4.3
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.168.0
Pillow==11.2.1
requests==2.32.3
python-multipart==0.0.20
fastapi==0.115.12
uvicorn==0.34.2
python-jose==3.3.0
python-magic==0.4.27
pydantic==2.11.3 